+++ src/app/workspace/organization/users/groups/organization-user-groups.component.scss
@@ -1,0 +1,17 @@
+@import '../../../../shared/variables';
+
+:host {
+  display: table-row;
+  vertical-align: middle;
+}
+
+.custom-control-label:after {
+  background-size: 70% 70%;
+}
+
+.vertical-line {
+  background-color: $success;
+  margin-left: 7px;
+  height: 35px;
+  width: 2px;
+}
