+++ src/app/workspace/organization/users/header/organization-users-header.component.scss
@@ -1,0 +1,9 @@
+@import '../../../../shared/variables';
+
+:host {
+  display: block;
+}
+
+.large-text {
+  font-size: 22px;
+}
