+++ src/app/workspace/organization/users/list/organization-user-list.component.scss
@@ -1,0 +1,9 @@
+@import '../../../../shared/variables';
+
+:host {
+  display: block;
+}
+
+.medium-text {
+  font-size: 20px;
+}
