+++ src/app/workspace/organization/users/organization-users.component.scss
@@ -5,25 +5,6 @@
   padding: 1em;
 }
 
-.filter {
-  width: 280px;
-}
-
-.large-text {
-  font-size: 22px;
-}
-
-.medium-text {
-  font-size: 20px;
-}
-
-.custom-control-label:after {
-  background-size: 70% 70%;
-}
-
-.vertical-line {
-  background-color: $success;
-  margin-left: 7px;
-  height: 35px;
-  width: 2px;
+.huge-text {
+  font-size: 38px;
 }
