+++ src/app/workspace/organization/users/organization-users.component.ts
@@ -25,13 +25,7 @@ import {selectOrganizationFromUrl} from '../../../core/store/organizations/organ
 import {OrganizationModel} from '../../../core/store/organizations/organization.model';
 import {UserModel} from '../../../core/store/users/user.model';
 import {selectAllUsers} from '../../../core/store/users/users.state';
-import {UsersAction} from '../../../core/store/users/users.action';
-import {PermissionModel} from '../../../core/store/permissions/permissions.model';
-import {Role} from '../../../shared/permissions/role';
-import {OrganizationsAction} from '../../../core/store/organizations/organizations.action';
 import {tap} from 'rxjs/operators';
-import {GroupModel} from '../../../core/store/groups/group.model';
-import {selectAllGroups} from '../../../core/store/groups/groups.state';
 
 @Component({
   templateUrl: './organization-users.component.html',
@@ -41,10 +35,6 @@ export class OrganizationUsersComponent implements OnInit {
 
   public organization$: Observable<OrganizationModel>;
   public users$: Observable<UserModel[]>;
-  public groups$: Observable<GroupModel[]>;
-
-  public expanded: { [email: string]: boolean } = {};
-  public newGroupName: { [email: string]: string } = {};
 
   constructor(private store: Store<AppState>) {
   }
@@ -52,106 +42,10 @@ export class OrganizationUsersComponent implements OnInit {
   public ngOnInit(): void {
     this.organization$ = this.store.select(selectOrganizationFromUrl);
     this.users$ = this.store.select(selectAllUsers).pipe(tap(this.sortUsers));
-    this.groups$ = this.store.select(selectAllGroups);
   }
 
   private sortUsers(users: UserModel[]): void {
     users.sort((user1, user2) => user1.name.localeCompare(user2.name));
-  }
-
-  public hasWritePermission(user: UserModel, organization: OrganizationModel): boolean {
-    const permissions = this.userPermissions(user, organization);
-    return permissions.roles.includes(Role.Write);
-  }
-
-  public hasManagePermission(user: UserModel, organization: OrganizationModel): boolean {
-    const permissions = this.userPermissions(user, organization);
-    return permissions.roles.includes(Role.Manage);
-  }
-
-  private userPermissions(user: UserModel, organization: OrganizationModel): PermissionModel {
-    let permissions = organization.permissions.users.find(userPermissions => {
-      return userPermissions.id === user.email;
-    });
-
-    if (permissions === undefined) {
-      permissions = this.addUserPermissions(user, organization);
-    }
-
-    return permissions;
-  }
-
-  private addUserPermissions(user: UserModel, organization: OrganizationModel): PermissionModel {
-    const createdPermissions: PermissionModel = {
-      id: user.email,
-      roles: []
-    };
-
-    organization.permissions.users.push(createdPermissions);
-    this.store.dispatch(new OrganizationsAction.Update({organization: organization}));
-
-    return createdPermissions;
-  }
-
-  public changeWritePermission(user: UserModel, organization: OrganizationModel): void {
-    if (this.hasWritePermission(user, organization)) {
-      this.removePermission(user, organization, Role.Write);
-    } else {
-      this.addPermission(user, organization, Role.Write);
-    }
-
-    this.store.dispatch(new OrganizationsAction.Update({organization: organization}));
-  }
-
-  public changeManagePermission(user: UserModel, organization: OrganizationModel): void {
-    if (this.hasManagePermission(user, organization)) {
-      this.removePermission(user, organization, Role.Manage);
-    } else {
-      this.addPermission(user, organization, Role.Manage);
-    }
-
-    this.store.dispatch(new OrganizationsAction.Update({organization: organization}));
-  }
-
-  private addPermission(user: UserModel, organization: OrganizationModel, addedRole: Role): void {
-    const permissions = this.userPermissions(user, organization);
-    permissions.roles.push(addedRole);
-  }
-
-  private removePermission(user: UserModel, organization: OrganizationModel, removedRole: Role): void {
-    const permissions = this.userPermissions(user, organization);
-    permissions.roles = permissions.roles.filter(role => role !== removedRole);
-  }
-
-  public addUser(): void {
-    const user: UserModel = {
-      email: '',
-      name: '',
-    };
-
-    this.store.dispatch(new UsersAction.Create({user: user}));
-  }
-
-  public updateUserName(user: UserModel): void {
-    this.store.dispatch(new UsersAction.Update({user: user}));
-  }
-
-  public updateUserEmail(user: UserModel): void {
-    this.store.dispatch(new UsersAction.Update({user: user}));
-  }
-
-  public blockUser(user: UserModel): void {
-    user.blocked = !user.blocked;
-
-    this.store.dispatch(new UsersAction.Update({user: user}));
-  }
-
-  public removeUser(user: UserModel): void {
-    this.store.dispatch(new UsersAction.Delete({userId: user.id}));
-  }
-
-  public userGroups(user: UserModel, groups: GroupModel[]): GroupModel[] {
-    return groups.filter(group => group.users && group.users.includes(user.email));
   }
 
 }
