+++ src/app/workspace/organization/users/user-filter/organization-user-filter.component.html
@@ -1,0 +1,11 @@
+<div class="form-control d-flex p-0 m-2 filter">
+  <input class="border-0 flex-grow-2 px-2" placeholder="Filter users"/>
+
+  <button class="btn search-button"
+          type="button"
+          title="Filter"
+          i18n-title="@@organizations.users.filter">
+    <i class="fa fa-search" aria-hidden="true"></i>
+    <span class="d-none" i18n="@@organizations.users.filter.button">Filter</span>
+  </button>
+</div>
