+++ src/app/workspace/organization/users/user-filter/organization-user-filter.component.scss
@@ -1,0 +1,9 @@
+@import '../../../../shared/variables';
+
+:host {
+  display: block;
+}
+
+.filter {
+  width: 280px;
+}
