+++ src/app/workspace/organization/users/user/organization-user.component.html
@@ -1,0 +1,67 @@
+<td class="px-1">
+  <i class="fa fa-caret-right large-text text-success {{ expanded ? 'fa-rotate-90' : null }}"
+     aria-hidden="true"
+     title="{{ expanded ? 'Close' : 'Open' }}"
+     i18n-title="@@organizations.user.expand"
+     (click)="expandedChange.emit()"></i>
+</td>
+
+<td class="px-1">
+  <input class="form-control edit-border"
+         type="text"
+         tabindex="1"
+         placeholder="Username"
+         [(ngModel)]="user.name"
+         (blur)="updateUserName(user)">
+</td>
+
+<td class="px-1">
+  <input class="form-control edit-border"
+         type="email"
+         tabindex="1"
+         placeholder="User email"
+         [ngModel]="user.email"
+         readonly>
+</td>
+
+<td class="px-1 text-center">
+  <div class="custom-control custom-control-inline custom-checkbox">
+    <input type="checkbox"
+           class="custom-control-input"
+           id="Create_proj_{{ user.id }}"
+           [ngModel]="hasWritePermission(user, organization)"
+           (change)="changeWritePermission(user, organization)">
+    <label class="custom-control-label"
+           for="Create_proj_{{ user.id }}">
+    </label>
+  </div>
+</td>
+
+<td class="px-1 text-center">
+  <div class="custom-control custom-control-inline custom-checkbox">
+    <input type="checkbox"
+           class="custom-control-input"
+           id="Manage_org_{{ user.id }}"
+           [ngModel]="hasManagePermission(user, organization)"
+           (change)="changeManagePermission(user, organization)">
+    <label class="custom-control-label"
+           for="Manage_org_{{ user.id }}">
+    </label>
+  </div>
+</td>
+
+<td class="px-1">
+  <i class="fas fa-fw {{ user.blocked ? 'fa-volume-mute' : 'fa-volume-up' }} text-danger mx-2 clickable"
+     aria-hidden="true"
+     title="{{ user.blocked ? 'Unblock' : 'Temporarily Block' }} User"
+     i18n-title="@@organizations.users.block"
+     (click)="blockUser(user)"></i>
+</td>
+
+<td class="px-1">
+  <i class="fas fa-fw fa-trash text-danger mx-2 clickable"
+     aria-hidden="true"
+     title="Remove User"
+     i18n-title="@@organizations.users.remove"
+     (click)="removeUser(user)"></i>
+</td>
