+++ src/app/workspace/organization/users/user/organization-user.component.scss
@@ -1,0 +1,20 @@
+@import '../../../../shared/variables';
+
+:host {
+  display: table-row;
+}
+
+.large-text {
+  font-size: 22px;
+}
+
+.custom-control-label:after {
+  background-size: 70% 70%;
+}
+
+.vertical-line {
+  background-color: $success;
+  margin-left: 7px;
+  height: 35px;
+  width: 2px;
+}
