+++ src/app/workspace/workspace.module.ts
@@ -34,6 +34,11 @@ import {WorkspaceService} from './workspace.service';
 import {OrganizationUsersComponent} from './organization/users/organization-users.component';
 import {OrganizationGroupsComponent} from './organization/groups/organization-groups.component';
 import {OrganizationRegistrationComponent} from './organization/registration/organization-registration.component';
+import {OrganizationUserListComponent} from './organization/users/list/organization-user-list.component';
+import {OrganizationUserComponent} from './organization/users/user/organization-user.component';
+import {OrganizationUsersHeaderComponent} from './organization/users/header/organization-users-header.component';
+import {OrganizationUserFilterComponent} from './organization/users/user-filter/organization-user-filter.component';
+import {OrganizationUserGroupsComponent} from './organization/users/groups/organization-user-groups.component';
 
 @NgModule({
   imports: [
@@ -45,6 +50,11 @@ import {OrganizationRegistrationComponent} from './organization/registration/org
   declarations: [
     OrganizationPermissionsComponent,
     OrganizationUsersComponent,
+    OrganizationUserListComponent,
+    OrganizationUserFilterComponent,
+    OrganizationUserGroupsComponent,
+    OrganizationUserComponent,
+    OrganizationUsersHeaderComponent,
     OrganizationGroupsComponent,
     OrganizationRegistrationComponent,
     ProjectPermissionsComponent,
